Unpack: Don't delete files if there was a warning or error

+++ plugins/unpack/unrar_dir.sh
@@ -12,13 +12,13 @@ process_directory()
 {
 	"$1" x -ai -c- -kb -o+ -p- -y -v -- "$2." "$3"
 	last=$?
-	[ $last -le 1 ] && ret=$last
+	[ $last -ge 1 ] && ret=$last
 	for fn in "$2"* ; do
 		if [ -d "${fn}" ] && [ ! -L "${fn}" ] ; then
 			name=$(basename "${fn}")
 			process_directory "$1" "${fn}/" "$3${name}/"
 			last=$?
-			[ $last -gt 1 ] && ret=$last
+			[ $last -ge 1 ] && ret=$last
 		fi
 	done
 	return $ret
@@ -32,8 +32,7 @@ else
 	ret=$?
 fi
 
-ret=$?
-if [ $ret -le 1 ] && [ "$5" != '' ] ; then
+if [ $ret -eq 0 ] && [ "$5" != '' ] ; then
 	OIFS=$IFS
 	IFS=';'
 	for file in "$5"
@@ -46,6 +45,6 @@ fi
 if [ "$6" != '' ] ; then
 	mkdir -p "$3"
 	mv "$6"* "$3"
-	rm -r "$6"
+	[ $? -eq 0 ] && rm -r "$6"
 fi
